requirements.txt
# Core ML & NLP
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2

# Data
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0

# Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.0

# Caching
redis==5.0.0

# Experiment tracking
mlflow==2.7.0

# File handling
python-docx==0.8.11
python-pptx==0.6.21
PyPDF2==3.0.1
Pillow==10.0.0
aiofiles==23.2.1

# Utils
python-dotenv==1.0.0
pyyaml==6.0.0
requests==2.31.0
httpx==0.24.0
loguru==0.7.2

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting / metrics
slowapi==0.1.9
prometheus-client==0.19.0

# Task queue / background jobs
celery==5.3.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# Dev tools
black==23.10.0
flake8==6.1.0
isort==5.12.0
mypy==1.5.0

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.0
ipywidgets==8.1.0

reportlab==4.0.8
python-docx==0.8.11
PyPDF2==3.0.1
